add diffusion method.

+++ compile_and_run.sh
@@ -1,5 +1,7 @@
 cd build
-rm result.ppm
+if test -e "result.ppm"; then
+    rm result.ppm
+fi
 cmake .. -GNinja
 ninja
-./ray-tracing >> result.ppm
+./ray-tracing >>result.ppm
